tiny fix
- [M] delete useless
- [A] test icp performance
